+++ pleo-antaeus-core/src/main/kotlin/io/pleo/antaeus/core/helpers/PaymentProcessingInput.kt
@@ -7,5 +7,5 @@ data class PaymentProcessingInput(
         val invoiceToProcess: Invoice,
         val paymentProvider: PaymentProvider,
         val maxNumberOfPaymentRetries: Int,
-        val paymentRetryDelayMs: Int
+        val paymentRetryDelayMs: Long
 )
